+++ src/main/java/com/montola/school/course/dto/QuizResponseDto.java
@@ -2,6 +2,7 @@ package com.montola.school.course.dto;
 
 import com.montola.school.course.enums.QuizType;
 import lombok.*;
+import java.math.BigDecimal;
 
 /**
  * @author avidewan
@@ -20,7 +21,7 @@ public class QuizResponseDto {
     private String instruction;
     private Integer timeLimit;
     private Integer totalMarks;
-    private Double passPercentage;
+    private BigDecimal passPercentage;
     private Long topicId;
     private String topicTitle;
     private int orderIndex;
